# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CorridorOps",
    "description": "CorridorOps - AI-Powered Automatic Block Planning System for Indian Railways (SIH PS #26027)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/defects/tms": {
      "post": {
        "tags": [
          "Defects & Department Requests"
        ],
        "summary": "Create Tms Defect",
        "operationId": "create_tms_defect_api_v1_defects_tms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TMSDefectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_TMSDefectSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/defects/smms": {
      "post": {
        "tags": [
          "Defects & Department Requests"
        ],
        "summary": "Create Smms Fault",
        "operationId": "create_smms_fault_api_v1_defects_smms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SMMSFaultCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_SMMSFaultSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/defects/tdms": {
      "post": {
        "tags": [
          "Defects & Department Requests"
        ],
        "summary": "Create Tdms Fault",
        "operationId": "create_tdms_fault_api_v1_defects_tdms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TDMSFaultCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_TDMSFaultSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/bdms/requests": {
      "post": {
        "tags": [
          "Defects & Department Requests"
        ],
        "summary": "Create Bdms Request",
        "operationId": "create_bdms_request_api_v1_bdms_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BDMSRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_BDMSRequestSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/defects": {
      "get": {
        "tags": [
          "Defects & Department Requests"
        ],
        "summary": "List Unified Defects",
        "operationId": "list_unified_defects_api_v1_defects_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "department",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Department"
            }
          },
          {
            "name": "section_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_UnifiedDefectItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/synthetic/generate": {
      "post": {
        "tags": [
          "Defects & Department Requests"
        ],
        "summary": "Trigger Synthetic Generation",
        "operationId": "trigger_synthetic_generation_api_v1_admin_synthetic_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyntheticGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BDMSRequestCreate": {
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "requested_window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Requested Window Start"
          },
          "requested_window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Requested Window End"
          },
          "crew_required": {
            "type": "integer",
            "title": "Crew Required",
            "default": 1
          },
          "machine_required": {
            "type": "boolean",
            "title": "Machine Required",
            "default": false
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "pending"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "department",
          "section_id",
          "requested_window_start",
          "requested_window_end"
        ],
        "title": "BDMSRequestCreate"
      },
      "BDMSRequestSchema": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "requested_window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Requested Window Start"
          },
          "requested_window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Requested Window End"
          },
          "crew_required": {
            "type": "integer",
            "title": "Crew Required",
            "default": 1
          },
          "machine_required": {
            "type": "boolean",
            "title": "Machine Required",
            "default": false
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "pending"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "department",
          "section_id",
          "requested_window_start",
          "requested_window_end",
          "id"
        ],
        "title": "BDMSRequestSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SMMSFaultCreate": {
        "properties": {
          "fault_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fault Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "fault_type": {
            "type": "string",
            "title": "Fault Type"
          },
          "severity": {
            "type": "integer",
            "title": "Severity"
          },
          "reported_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Date"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "section_id",
          "fault_type",
          "severity"
        ],
        "title": "SMMSFaultCreate"
      },
      "SMMSFaultSchema": {
        "properties": {
          "fault_id": {
            "type": "string",
            "title": "Fault Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "fault_type": {
            "type": "string",
            "title": "Fault Type"
          },
          "severity": {
            "type": "integer",
            "title": "Severity"
          },
          "reported_date": {
            "type": "string",
            "format": "date-time",
            "title": "Reported Date"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "fault_id",
          "section_id",
          "fault_type",
          "severity",
          "reported_date",
          "id"
        ],
        "title": "SMMSFaultSchema"
      },
...
      "StandardResponse_BDMSRequestSchema_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BDMSRequestSchema"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[BDMSRequestSchema]"
      },
...
      "StandardResponse_List_UnifiedDefectItem__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/UnifiedDefectItem"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[List[UnifiedDefectItem]]"
      },
...
      "StandardResponse_SMMSFaultSchema_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SMMSFaultSchema"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[SMMSFaultSchema]"
      },
...
      "StandardResponse_TDMSFaultSchema_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TDMSFaultSchema"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[TDMSFaultSchema]"
      },
      "StandardResponse_TMSDefectSchema_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TMSDefectSchema"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[TMSDefectSchema]"
      },
...
      "StandardResponse_dict_": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[dict]"
      },
...
      "SyntheticGenerateRequest": {
        "properties": {
          "table": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "config_overrides": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Overrides"
          }
        },
        "type": "object",
        "title": "SyntheticGenerateRequest"
      },
      "TDMSFaultCreate": {
        "properties": {
          "fault_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fault Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "fault_type": {
            "type": "string",
            "title": "Fault Type"
          },
          "severity": {
            "type": "integer",
            "title": "Severity"
          },
          "reported_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Date"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "section_id",
          "fault_type",
          "severity"
        ],
        "title": "TDMSFaultCreate"
      },
      "TDMSFaultSchema": {
        "properties": {
          "fault_id": {
            "type": "string",
            "title": "Fault Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "fault_type": {
            "type": "string",
            "title": "Fault Type"
          },
          "severity": {
            "type": "integer",
            "title": "Severity"
          },
          "reported_date": {
            "type": "string",
            "format": "date-time",
            "title": "Reported Date"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "fault_id",
          "section_id",
          "fault_type",
          "severity",
          "reported_date",
          "id"
        ],
        "title": "TDMSFaultSchema"
      },
      "TMSDefectCreate": {
        "properties": {
          "defect_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defect Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "defect_type": {
            "type": "string",
            "title": "Defect Type"
          },
          "severity": {
            "type": "integer",
            "title": "Severity"
          },
          "reported_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Date"
          },
          "speed_restriction_kmph": {
            "type": "number",
            "title": "Speed Restriction Kmph",
            "default": 0.0
          },
          "overdue_flag": {
            "type": "boolean",
            "title": "Overdue Flag",
            "default": false
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "section_id",
          "defect_type",
          "severity"
        ],
        "title": "TMSDefectCreate"
      },
      "TMSDefectSchema": {
        "properties": {
          "defect_id": {
            "type": "string",
            "title": "Defect Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "defect_type": {
            "type": "string",
            "title": "Defect Type"
          },
          "severity": {
            "type": "integer",
            "title": "Severity"
          },
          "reported_date": {
            "type": "string",
            "format": "date-time",
            "title": "Reported Date"
          },
          "speed_restriction_kmph": {
            "type": "number",
            "title": "Speed Restriction Kmph",
            "default": 0.0
          },
          "overdue_flag": {
            "type": "boolean",
            "title": "Overdue Flag",
            "default": false
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "defect_id",
          "section_id",
          "defect_type",
          "severity",
          "reported_date",
          "id"
        ],
        "title": "TMSDefectSchema"
      },
...
      "UnifiedDefectItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_table": {
            "type": "string",
            "title": "Source Table"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "defect_or_fault_type": {
            "type": "string",
            "title": "Defect Or Fault Type"
          },
          "severity": {
            "type": "integer",
            "title": "Severity"
          },
          "reported_date": {
            "type": "string",
            "format": "date-time",
            "title": "Reported Date"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_table",
          "section_id",
          "department",
          "defect_or_fault_type",
          "severity",
          "reported_date",
          "is_simulated"
        ],
        "title": "UnifiedDefectItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}